feat: adds container for data service coordination
the SourceDataServices class can be used to encapsulate the knowledge of
the different services and coordination of their reads. To over-engineer
a solution you could utilize some reflection processing that
automatically finds and creates all the services and then this service
can refresh them as needed. This would allow new services to be added by
just creating a new class without having to modify something elsewhere